+++ core/shared/src/main/scala/cats/effect/concurrent/Ref.scala
@@ -34,7 +34,7 @@ import scala.annotation.tailrec
  * An asynchronous, concurrent mutable reference.
  *
  * Provides safe concurrent access and modification of its content, but no
- * functionality for synchronisation, which is instead handled by [[Promise]].
+ * functionality for synchronisation, which is instead handled by [[Pledge]].
  * For this reason, a `Ref` is always initialised to a value.
  *
  * The default implementation is nonblocking and lightweight, consisting essentially
