+++ laws/shared/src/main/scala/cats/effect/laws/ConcurrentEffectLaws.scala
@@ -18,7 +18,7 @@ package cats
 package effect
 package laws
 
-import cats.effect.laws.util.Pledge
+import cats.effect.concurrent.Pledge
 import cats.implicits._
 import cats.laws._
 
@@ -43,14 +43,14 @@ trait ConcurrentEffectLaws[F[_]] extends ConcurrentLaws[F] with EffectLaws[F] {
   def runCancelableStartCancelCoherence[A](a: A) = {
     // Cancellation via runCancelable
     val f1 = Pledge[IO, A].flatMap { effect1 =>
-      val never = F.cancelable[A](_ => effect1.complete[IO](a))
-      F.runCancelable(never)(_ => IO.unit).flatten *> effect1.await[IO]
+      val never = F.cancelable[A](_ => effect1.complete(a))
+      F.runCancelable(never)(_ => IO.unit).flatten *> effect1.get
     }
     // Cancellation via start.flatMap(_.cancel)
     val f2 = Pledge[IO, A].flatMap { effect2 =>
-      val never = F.cancelable[A](_ => effect2.complete[IO](a))
+      val never = F.cancelable[A](_ => effect2.complete(a))
       val task = F.start(never).flatMap(_.cancel)
-      F.runAsync(task)(_ => IO.unit) *> effect2.await[IO]
+      F.runAsync(task)(_ => IO.unit) *> effect2.get
     }
     f1 <-> f2
   }
