--- laws/shared/src/test/scala/cats/effect/PledgeTests.scala
@@ -1,61 +1,0 @@
-/*
- * Copyright (c) 2017-2018 The Typelevel Cats-effect Project Developers
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package cats.effect
-
-import cats.effect.laws.util.Pledge
-import scala.util.Success
-
-class PledgeTests extends BaseTestsSuite {
-  testAsync("happy path") { implicit ec =>
-    val io = for {
-      p <- Pledge[IO, Int]
-      fiber <- p.await[IO].start
-      _ <- p.complete[IO](10)
-      r <- fiber.join
-    } yield r
-
-    val f = io.unsafeToFuture()
-    ec.tick()
-    f.value shouldBe Some(Success(10))
-  }
-
-  testAsync("cannot complete twice") { implicit ec =>
-    val io = for {
-      p <- Pledge[IO, Int]
-      _ <- p.complete[IO](10)
-      _ <- p.complete[IO](10)
-    } yield ()
-
-    val f = io.unsafeToFuture()
-    ec.tick()
-
-    intercept[IllegalStateException] {
-      f.value.get.get
-    }
-  }
-
-  testAsync("non-completed pledge is non-terminating") { implicit ec =>
-    val io = for {
-      p <- Pledge[IO, Int]
-      r <- p.await[IO]
-    } yield r
-
-    val f = io.unsafeToFuture()
-    ec.tick()
-    f.value shouldBe None
-  }
-}
